Refactor project to remove CamelNaming from file names Fixes #43
