+++ src/wiseagents/llm/openai_API_wise_agent_LLM.py
@@ -1,4 +1,4 @@
-from wiseagents.llm.WiseAgentRemoteLLM import WiseAgentRemoteLLM
+from wiseagents.llm.wise_agent_remote_LLM import WiseAgentRemoteLLM
 import requests
 import openai
