+++ tests/wiseagents/llm/test_WiseAgentRemoteLLM.py
@@ -1,5 +1,5 @@
-from wiseagents.llm.LangChainWiseAgentRemoteLLM import LangChainWiseAgentRemoteLLM
-from wiseagents.llm.OpenaiAPIWiseAgentLLM import OpenaiAPIWiseAgentLLM
+from wiseagents.llm.lang_chain_wise_agent_remote_LLM import LangChainWiseAgentRemoteLLM
+from wiseagents.llm.openai_API_wise_agent_LLM import OpenaiAPIWiseAgentLLM
 
   
 def test_langchain():
